Change force_fps and force_ofps to double
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@25997 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_nemesi.c
@@ -108,7 +108,7 @@ static void link_session_and_fetch_conf(Nemesi_DemuxerStreamData * ndsd,
                                         rtp_session * sess,
                                         rtp_buff * buff, unsigned int * fps)
 {
-    extern float force_fps;
+    extern double force_fps;
     rtp_ssrc *ssrc = NULL;
     rtp_frame * fr = &ndsd->first_pkt[stype];
     rtp_buff trash_buff;
